🔧 refactor: Remove useCurrentPromptData hook and clean up related components; enhance PromptVersions layout
