+++ client/src/hooks/Prompts/index.ts
@@ -1,3 +1,2 @@
 export { default as useCategories } from './useCategories';
 export { default as usePromptGroupsNav } from './usePromptGroupsNav';
-export { useCurrentPromptData } from './useCurrentPromptData';
